fix test warnings
