+++ mix.exs
@@ -9,6 +9,7 @@ defmodule PaperTrail.Mixfile do
       description: description(),
       build_embedded: Mix.env == :prod,
       start_permanent: Mix.env == :prod,
+      elixirc_paths: elixirc_paths(Mix.env),
       package: package(),
       deps: deps()
     ]
@@ -50,4 +51,6 @@ defmodule PaperTrail.Mixfile do
       }
     ]
   end
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 end
