+++ test/support/repo.ex
@@ -1,0 +1,22 @@
+defmodule PaperTrail.Repo do
+  use Ecto.Repo, otp_app: :paper_trail
+end
+
+defmodule User do
+  use Ecto.Schema
+
+  import Ecto.Changeset
+
+  schema "users" do
+    field :token, :string
+    field :username, :string
+
+    timestamps()
+  end
+
+  def changeset(model, params \\ %{}) do
+    model
+    |> cast(params, [:token, :username])
+    |> validate_required([:token, :username])
+  end
+end
